requirements.txt
boto==2.49.0
botocore==1.20.54
boto3==1.17.54
apache-airflow==2.0.2
pathlib_mate==1.0.1
pynamodb==5.1.0
